% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/observed-data.r
\docType{data}
\name{observed}
\alias{observed}
\title{Number of syndromic cases per time unit in the matrix format}
\format{a \code{matrix} .}
\source{
package \code{vetsyn}
}
\usage{
observed
}
\description{
This dataset exemplifies the case when a user of the package
{vetsyn} may already have data cleaned and organized in a matrix format,
with the columns representing counts for each syndromic group monitored,
and the rows representing the time units (which could be daily, weekly, etc).
This could be the case when users prefer not to convert raw data to a matrix
format using the functionalities of the package, but instead their
own customized functions and adjustments. The matrix could then be readly
transformed into a syndromic object using the function \code{syndromic()}.
}
\author{
National Veterinary Institute of Sweden (SVA), 2013
}

