\docType{methods}
\name{plot-methods}
\alias{plot}
\alias{plot-methods}
\alias{plot,syndromic-method}
\title{Plot}
\description{
  Plot a \code{\link{syndromic}} object.
}
\section{Methods}{
  \describe{ \item{\code{signature(object = "syndromic")}}{
  Show information for the syndromic object. } }
}
\keyword{methods}

