\docType{methods}
\name{show}
\alias{show}
\title{Show}
\description{
  Shows information for a syndromic object
  (\code{syndromicD} or \code{syndromicW}).
}
\section{Methods}{
  \describe{ \item{\code{signature(object = "syndromic")}}{
  Show information for a syndromic object. } }
}

\section{Methods}{
  \describe{ \item{\code{signature(object = "syndromic")}}{
  Show information for a syndromic object. } }
}
\keyword{methods}

