% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syndromicW.r
\docType{class}
\name{syndromicW-class}
\alias{syndromicW-class}
\title{Class \code{"syndromicW"}}
\description{
Syndromic is the main class of the \code{vetsyn} package, but it has been 
specified for different types of data streams. \code{syndromicD} is the class
used for data that are recorded and monitored DAILY, while \code{syndromicW} is
used when the time points for monitoring refer to WEEKS.
}
\details{
The \code{syndromicD} or \code{syndromicW} classes store observed data in 
a format appropriate
for monitoring, and contain several slots to store input and outputs of 
analysis (temporal monitoring).
Functions are available to create an object of the class syndromic from data 
already cleaned and prepared for monitoring, or alternatively from raw observed data.
}
\examples{
## Load data
data(observedW)
my.syndromicW <- syndromicW(observed,min.week=1, min.year=2011, 
                              max.week=22, max.year=2013)

}
