\docType{methods}
\name{plot-methods}
\alias{plot-methods}
\title{Plot}
\description{
  Plot a syndromic (\code{syndromicD} or \code{syndromicW})
  object.

  Plot a syndromic (\code{syndromicD} or \code{syndromicW})
  object.
}
\section{Methods}{
  \describe{ \item{\code{signature(object =
  "syndromicD")}}{ Show information for the syndromic
  object. } }
}

\section{Methods}{
  \describe{ \item{\code{signature(object =
  "syndromicW")}}{ Show information for the syndromic
  object. } }
}
\keyword{methods}

