% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r, R/plotW.R
\docType{methods}
\name{plot-methods}
\alias{plot-methods}
\title{Plot}
\usage{
\S4method{plot}{syndromicD}(x, syndromes = NULL, window = 365,
  baseline = FALSE, UCL = 1, algorithms = NULL, limit = NULL)

\S4method{plot}{syndromicW}(x, syndromes = NULL, window = 52,
  baseline = FALSE, UCL = 1, algorithms = NULL, limit = NULL)
}
\arguments{
\item{x}{a syndromic object}

\item{syndromes}{an optional parameter, if not specified, all
columns in the slot observed (or baseline of the
 \code{syndromic} object
will be used. The user can choose to restrict the plotting to 
a few syndromic groups listing their name or column position
in the observed matrix. See examples.}

\item{window}{the number of time points to plot, always finishing at the 
last time point recorded.}

\item{baseline}{whether to plot the baseline, by default equal to TRUE.}

\item{UCL}{the dimension of the slot UCL, from the syndromic object, from which
the user wants to plot the UCL. Set to NULL or to 0 if it is not desired to plot the UCL.}

\item{algorithms}{an optional parameter specifying which dimensions
of the alarm slot to plot. If not specified (NULL), all are plotted. If set 
to zero, none are plotted.}

\item{limit}{an optional parameter establishing a score above which
alarms are considered worth of notice. Notice that this is not a
statistical value, but a score equivalent to the number of algorithms
employed and the number of limits set to them. See the tutorial for details.}

\item{x}{a syndromic object}

\item{syndromes}{an optional parameter, if not specified, all
columns in the slot observed (or baseline of the
 \code{syndromic} object
will be used. The user can choose to restrict the plotting to 
a few syndromic groups listing their name or column position
in the observed matrix. See examples.}

\item{window}{the number of time points to plot, always finishing at the 
last time point recorded.}

\item{baseline}{whether to plot the baseline, by default equal to TRUE.}

\item{UCL}{the dimension of the slot UCL, from the syndromic object, from which
the user wants to plot the UCL. Set to NULL or to 0 if it is not desired to plot the UCL.}

\item{algorithms}{an optional parameter specifying which dimensions
of the alarm slot to plot. If not specified (NULL), all are plotted. If set 
to zero, none are plotted.}

\item{limit}{an optional parameter establishing a score above which
alarms are considered worth of notice. Notice that this is not a
statistical value, but a score equivalent to the number of algorithms
employed and the number of limits set to them. See the tutorial for details.}
}
\description{
Plot a syndromic (\code{syndromicD} or \code{syndromicW}) object.

Plot a syndromic (\code{syndromicD} or \code{syndromicW}) object.
}
\section{Methods}{
 \describe{
\item{\code{signature(object = "syndromicD")}}{
Show information for the syndromic object.
}
}

 \describe{
\item{\code{signature(object = "syndromicW")}}{
Show information for the syndromic object.
}
}
}
\keyword{methods}

