% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.r, R/plotW.R
\docType{methods}
\name{plot-methods}
\alias{plot-methods}
\title{Plot}
\usage{
\S4method{plot}{syndromicD}(x, syndromes = NULL, window = 365,
  baseline = FALSE, UCL = 1, algorithms = NULL, limit = NULL)

\S4method{plot}{syndromicW}(x, syndromes = NULL, window = 52,
  baseline = FALSE, UCL = 1, algorithms = NULL, limit = NULL)
}
\description{
Plot a syndromic (\code{syndromicD} or \code{syndromicW}) object.

Plot a syndromic (\code{syndromicD} or \code{syndromicW}) object.
}
\section{Methods}{
 \describe{
\item{\code{signature(object = "syndromicD")}}{
Show information for the syndromic object.
}
}
}

\section{Methods}{
 \describe{
\item{\code{signature(object = "syndromicW")}}{
Show information for the syndromic object.
}
}
}
\keyword{methods}

