\docType{methods}
\name{show-methods}
\alias{show}
\alias{show-methods}
\alias{show,syndromic-method}
\title{Show}
\description{
  Shows information for the \code{\link{syndromic}} object.
}
\section{Methods}{
  \describe{ \item{\code{signature(object = "syndromic")}}{
  Show information for the syndromic object. } }
}
\keyword{methods}

