% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.r
\docType{methods}
\name{show}
\alias{show}
\alias{show,syndromicW-method}
\title{Show}
\usage{
\S4method{show}{syndromicD}(object)

\S4method{show}{syndromicW}(object)
}
\arguments{
\item{object}{a syndromic object}
}
\description{
Shows information for a syndromic object (\code{syndromicD} or 
\code{syndromicW}).
}
\section{Methods}{
 \describe{
\item{\code{signature(object = "syndromic")}}{
Show information for a syndromic object.
}
}
}
\keyword{methods}

