% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show.r, R/showW.r
\docType{methods}
\name{show}
\alias{show}
\title{Show}
\usage{
\S4method{show}{syndromicD}(object)

\S4method{show}{syndromicW}(object)
}
\description{
Shows information for a syndromic object (\code{syndromicD} or
\code{syndromicW}).
}
\section{Methods}{
 \describe{
\item{\code{signature(object = "syndromic")}}{
Show information for a syndromic object.
}
}
}

\section{Methods}{
 \describe{
\item{\code{signature(object = "syndromic")}}{
Show information for a syndromic object.
}
}
}
\keyword{methods}

